added runtime error to handle the linoodle not having been built, symlink guards, and game detect

+++ lib/iostore.py
@@ -58,13 +58,30 @@ import config
 # to import this module and get a helpful error, not a raw ctypes OSError.
 _oodle = None
 
+# Third party dependency linoodle hardcodes the name oo2core_8_win64.dll
+LINOODLE_REQUIRED_DLL_NAME = "oo2core_8_win64.dll"
+
 
 def _linoodle_paths(dll_path):
     build_dir = os.path.join(
         os.path.dirname(os.path.dirname(os.path.abspath(__file__))),
         "third_party", "build")
     so_path = os.path.join(build_dir, "liblinoodle.so")
-    dll_link = os.path.join(build_dir, "oo2core_8_win64.dll")
+    if not os.path.isfile(so_path):
+        raise RuntimeError(
+            "linoodle has not been built,\n"
+            f"{so_path}\n"
+            "please make sure to run the following commands found in the README\n"
+            "   cmake -S third_party -B third_party/build -DCMAKE_BUILD_TYPE=Debug\n"
+            "   cmake --build third_party/build --target linoodle"
+        )
+    # Build the symlink to say that it is "oo2core_8_win64.dll",
+    # in reality we can put any version of the oodle dll that we have
+    dll_link = os.path.join(build_dir, LINOODLE_REQUIRED_DLL_NAME)
+    # os.path.lexists() is true for regular files as well as symlinks, so
+    # os.remove() would unlink a real oo2core_8_win64.dll a user had put there.
+    if os.path.lexists(dll_link) and not os.path.islink(dll_link):
+        return so_path, build_dir
     if not os.path.lexists(dll_link) or os.path.realpath(dll_link) != os.path.realpath(dll_path):
         if os.path.lexists(dll_link):
             os.remove(dll_link)
